requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
youtube-transcript-api==1.0.3
python-dotenv==1.0.1
gunicorn==21.2.0 